==> code/41_chla_satellite_vs_obsRmd.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: inline
---

This script aims to 
  * examine the correlation between DO and rs variables
  * check the multicollinearity among RS variables. 

# Set up

```{r Packages and dir, include=FALSE}
# To clear your environment
remove(list = ls())

library(readxl)
library(dplyr)
library(sf)
library(lubridate)
library(cowplot)

source('./code/_path.R')
```



# Data

## Sample data

  - The `xlsx` and `shp` data are the same after checking.

```{r - xlsx, warning=FALSE}

xls <- paste0('./data/Chlorophyll_a data/Chlorophyll_a_1243776172852889586.xlsx'); xls

df <- readxl::read_excel(path = xls)
names(df)


chl <- df %>%
  dplyr::select(-c(organization_formal_name, 
                   x, y, note, activity_start_time, activity_start_timezone, monitoring_location_name,
                   sample_fraction)) %>%
  dplyr::mutate(year = year(activity_start_date),
                month = month(activity_start_date)) %>%
  dplyr::filter(year >= 2000, 
                month >=5,
                month <=9,
                activity_depth_height >= 0,
                activity_depth_height <=5
                ) %>%
  arrange(monitoring_location_identifier) %>%
  as.data.frame()

hist(chl$activity_depth_height)
```


```{r - shp}
f <- './data/Chlorophyll_a data/Chlorophyll_a_7003234383276519096/Chlorophyll_Observation_Points.shp'; f
d <- st_read(f)

chl.sf <- d %>%
  dplyr::rename('id' = 'monitoring',
                'date' = 'activityst',
                'depth' = 'activityde',
                'depth_unit' = 'activity_3', 
                'var_name' = 'characteri', 
                'var_value' = 'resultmeas', 
                'var_unit' = 'resultme_1'
                ) %>%
  dplyr::select(-c(organizati, 
                   note, activity_1, activity_2, monitori_1, resultsamp)) %>%
  dplyr::mutate(year = year(date),
                month = month(date)) %>%
  dplyr::filter(year >= 2000, 
                month >=5,
                month <=9,
                depth >= 0,
                depth <= 5
                ) 

# Convert 3D geometries to 2D
chl.sf_2D <- chl.sf %>%
  st_drop_geometry() %>%
  st_as_sf(., coords = c("longitude", "latitude"), crs = 4326, remove = FALSE)


f <- './data/Chlorophyll_a data/Chlorophyll_a_7003234383276519096/Chlorophyll_Observation_Points_filter.shp'
# st_write(obj = chl.sf_2D, dsn = f, driver = "ESRI Shapefile", delete_dsn = T)
```


## GEE imagery

  - Upload `chl.sf_2D` to GEE, and 

  - Run GEE code `users/yingjieli/DZT/01b_ic2csv_chl_a`, output csv files
  
  - Load csv files in R

```{r}

dir_chla <- "./data/from_gee/Img2Table_chla/"

fs <- list.files(path = dir_chla, pattern = '_chlor_a.csv$', full.names = T)
# fs

# Initialize an empty list to store individual data frames
df_list <- list()

# Loop through each CSV file and read it into a data frame
for (f in fs) {
  # print(basename(f))
  sensor = gsub('_chlor_a.csv|[0-9]|_', '', basename(f))
  df <- readr::read_csv(f, show_col_types = FALSE) %>%
    dplyr::select(-`.geo`, -`system:index`, -id) %>%
    dplyr::select(date:year, everything()) %>%
    gather(key = date_s, value = chla_s, 12:ncol(.)) %>%
    dplyr::filter(chla_s != "No data") %>%
    mutate(date_s = as.Date(as.character(date_s), format = "%Y-%m-%d"),
           date = as.POSIXct(date/1000, origin = "1970-01-01", tz = "UTC"),
           date = as.Date(date, format = "%Y%m%d"),
           sensor = sensor,
           date_dif = date - date_s) %>%
    dplyr::filter(abs(date_dif) < 1) %>%
    as.data.frame()
  df_list <- append(df_list, list(df))
}

# 
# d <- df %>%
#   dplyr::filter(chla_s != "No data") %>%
#   mutate(date_s = as.Date(as.character(date_s), format = "%Y-%m-%d"),
#          date = as.POSIXct(date/1000, origin = "1970-01-01", tz = "UTC"),
#          date = as.Date(date, format = "%Y%m%d"),
#          date_dif = date - date_s) %>%
#   dplyr::filter(abs(date_dif) < 1) %>%
#   as.data.frame()

# Bind all data frames in the list into a single data frame
dfs <- do.call(rbind, df_list) %>%
  dplyr::filter(depth < 0.1,
                var_value > 0) %>%
  dplyr::mutate(chla_s = as.numeric(chla_s)) %>%
  group_by(sensor, date, year, featureID, latitude, longitude) %>%
  dplyr::summarise_at(vars(var_value, chla_s), ~ mean(., na.rm = TRUE))

```


# Plot

```{r - pre-process}
# Function to remove outliers using IQR
remove_outliers <- function(x, q1 = 0.15) {
  # Calculate Q1 (25th percentile) and Q3 (75th percentile)
  Q1 <- quantile(x, q1, na.rm = TRUE)
  Q3 <- quantile(x, 1-q1, na.rm = TRUE)
  
  # Calculate IQR
  IQR <- Q3 - Q1
  
  # Define the lower and upper bounds for outliers
  lower_bound <- Q1 - 1.5 * IQR
  upper_bound <- Q3 + 1.5 * IQR
  
  # Remove outliers
  x[x < lower_bound | x > upper_bound] <- NA # Mark outliers as NA
  
  return(x)
}


```



```{r - plot}

# Apply the remove_outliers function to each numeric column ------------------------------
sens <- 'aqua'
# sens <- 'terr'


dfs_ <- dfs %>%
  as.data.frame() %>%
  dplyr::filter(sensor == sens) %>%
  dplyr::mutate_at(vars(var_value, chla_s), remove_outliers) %>%
  drop_na()


library(ggpubr)

# Calculate the maximum x-value for each category
max_x_values <- dfs_ %>%
  group_by(sensor, year) %>%
  dplyr::summarize(max_x = max(var_value, na.rm = T),
                   max_y = max(chla_s, na.rm = T))

# Merge the max_x_values back into the original data frame
dfs_ <- dfs_ %>%
  left_join(max_x_values, by = c('sensor', "year"))


p <- dfs_ %>%
  ggplot(aes(x = var_value, y = chla_s)) +
  geom_point(aes(color = year), alpha = 0.5)+
  geom_abline(slope = 1, colour = 'red') +
  geom_smooth(method = "lm", se = TRUE, color = "blue", linetype = "dashed") +
  stat_regline_equation(
    label.x = max(dfs_$max_x)/1.3, label.y = max(dfs_$max_y)/2, aes(label = after_stat(eq.label))) + # Add equation
  stat_cor(
    label.x = max(dfs_$max_x)/1.3, label.y = max(dfs_$max_y)/2-2, aes(label = paste(..rr.label.., sep = "~~~"))) + # Add R² annotation
  xlab('Chlorophyll_a (ug/l) - Observed') + ylab('Chlorophyll_a (ug/l) - MODIS') +
  theme_bw()
p
f <- paste0(dir.fig, 'SM_chl_a_obs_vs_satellite_', sens, '_', today, '.png'); f
ggsave(filename = f, plot = last_plot(), width = 6, height = 5, units = 'in', dpi = 300)



# Fit a linear regression model
model <- lm(var_value ~ chla_s, data = dfs_)
model
# Make predictions using the fitted model
predictions <- predict(model, dfs_)
# Calculate the Mean Absolute Error
mae <- mean(abs(predictions - dfs_$var_value))
# Print the MAE
print(paste("Mean Absolute Error:", mae))

# 计算 var1 和 var2 之间的相关系数
cor_var1_var2 <- cor(dfs_$var_value, dfs_$chla_s)

# 打印结果
print(paste("var1 - var2 Correlation Coefficient:", cor_var1_var2))



## facet plot ----------------------------------------------------------------------------

p + facet_wrap(~year, scales = 'free_y')
f <- paste0(dir.fig, 'SM_chl_a_obs_vs_satellite_', sens, '_', today, '_.png'); f
ggsave(filename = f, plot = last_plot(), width = 6*2, height = 5*2, units = 'in', dpi = 300)
```
